backend/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.41.0
pydantic==2.12.5
python-dotenv==1.2.1
faiss-cpu==1.13.2
numpy==2.4.2
google-cloud-aiplatform==1.138.0